next.config.js를 esm 기반으로 변경
